cloud-function/api-gateway/requirements.txt
Flask==2.1.0
google-cloud-workflows==1.8.0
python-dotenv==0.21.0